Decal + Terrain support
Decal materials are set up the same as other materials.

Terrain only supports albedo input, this is due to editor limitations.
Because of this all terrain surfaces are assumed non-metal, have a
specular value that matches rock/dirt and have a set roughness, this
roughness is slightly influenced by the albedo to provide some
variation.