Added Emission Channel + diff/spec energy conservation
I planned on added energy conservation for the whole BRDF once I found a
quick solution. I found such a solution (1.0 - fresnel)*diffuse.

Emission now added as well. It is HDR enabled, so will go >1

I fixed a few bugs, and had accidently overwritten the default probe
with the optional one so that is now fixed.